+++ docs/changelog.rst
@@ -2,7 +2,8 @@
 Changelog
 =========
 
-* :feature:`11702` Zerox Base Swaps through the latest settler  will now be properly decoded.
+* :bug:`11709` History events page will now properly refresh after marking an asset as spam or ignoring it from the context menu.
+* :feature:`11702` Zerox Base Swaps through the latest settler will now be properly decoded.
 
 * :release:`1.42.0 <2025-02-20>`
 * :feature:`-` Octant v2 migration is now supported and GLM locked in v2 detected.